First draft of pediatricAgeGroup filter
