+++ app/pathogen/arbovirus/analyze/page.tsx
@@ -2,7 +2,7 @@
 
 import React, { useCallback, useContext, useMemo } from "react";
 import { Card, CardContent, CardHeader, CardTitle } from "@/components/ui/card";
-import { Filters } from "@/app/pathogen/arbovirus/dashboard/filters";
+import { FilterableField, Filters } from "@/app/pathogen/arbovirus/dashboard/filters";
 import { ArboDataTable } from "@/app/pathogen/arbovirus/analyze/ArboDataTable";
 import clsx from "clsx";
 import { CardConfiguration, CardStyle, CardType, getConfigurationForCard } from "@/components/ui/card-collection/card-collection-types";
@@ -10,7 +10,7 @@ import { CardCollection } from "@/components/ui/card-collection/card-collection"
 import { VisualizationId, addToVisualizationInformation } from "../visualizations/visualizations";
 import { ZoomIn } from "lucide-react";
 import { useRouter } from "next/navigation";
-import { ArboContext } from "@/contexts/arbo-context";
+import { ArboContext } from "@/contexts/arbo-context/arbo-context";
 import { useArboDataInsights } from "@/hooks/useArboDataInsights";
 
 const VisualizationCard = (props: {
