+++ app/pathogen/arbovirus/analyze/recharts/estimate-count-by-un-region-and-arbovirus-graph.tsx
@@ -5,7 +5,7 @@ import {
   arbovirusesSF,
   convertArboSFtoArbo,
 } from "../recharts";
-import { ArboContext } from "@/contexts/arbo-context";
+import { ArboContext } from "@/contexts/arbo-context/arbo-context";
 import {
   typedGroupBy,
   typedObjectEntries,
