+++ contexts/arbo-context/arbo-context.tsx
@@ -1,0 +1,134 @@
+"use client";
+
+import React, { createContext, useReducer, useEffect } from "react";
+import { MapProvider, MapRef, useMap } from "react-map-gl";
+import { QueryClient, QueryClientProvider } from "@tanstack/react-query";
+import { Hydrate as RQHydrate, HydrateProps } from "@tanstack/react-query";
+import { useArboData } from "@/hooks/useArboData";
+import { filterData } from "./filter-update-steps/apply-new-selected-filters-step";
+import { handleArboFilterUpdate } from "./arbo-filter-update-steps";
+
+export interface ArboContextType extends ArboStateType {
+  dispatch: React.Dispatch<ArboAction>;
+}
+
+export interface ArboStateType {
+  filteredData: any[];
+  selectedFilters: { [key: string]: string[] };
+  dataFiltered: boolean;
+}
+
+export interface ArboAction {
+  type: ArboActionType;
+  payload: any;
+}
+export enum ArboActionType {
+  UPDATE_FILTER = "UPDATE_FILTER",
+  INITIAL_DATA_FETCH = "INITIAL_DATA_FETCH",
+  ADD_FILTERS_TO_MAP = "ADD_FILTERS_TO_MAP",
+  RESET_FILTERS = "RESET_FILTERS",
+}
+
+export const initialState: ArboStateType = {
+  filteredData: [],
+  selectedFilters: {
+    ["pathogen"]: ["DENV", "ZIKV", "CHIKV", "YF", "WNV", "MAYV"],
+  },
+  dataFiltered: false,
+};
+
+export const arboReducer = (
+  state: ArboStateType,
+  action: ArboAction,
+  map: MapRef | undefined
+) => {
+  switch (action.type) {
+    case ArboActionType.UPDATE_FILTER:
+      return handleArboFilterUpdate({
+        state,
+        action,
+        map
+      }).state;
+    case ArboActionType.INITIAL_DATA_FETCH:
+      return {
+        ...state,
+        filteredData: action.payload.data,
+        selectedFilters: initialState.selectedFilters,
+        dataFiltered: false,
+      };
+    case ArboActionType.RESET_FILTERS:
+      return {
+        ...state,
+        filteredData: filterData(action.payload.data, {}),
+        selectedFilters: initialState.selectedFilters,
+        dataFiltered: false,
+      };
+
+    default:
+      return state;
+  }
+};
+
+export const ArboContext = createContext<ArboContextType>({
+  ...initialState,
+  dispatch: (obj) => {
+    console.debug("dispatch not initialized", obj);
+  },
+});
+
+export const ArboProviders = ({ children }: { children: React.ReactNode }) => {
+  const queryClient = new QueryClient({
+    defaultOptions: {
+      queries: {
+        staleTime: Infinity,
+        cacheTime: Infinity,
+      },
+    },
+  });
+
+  return (
+    <MapProvider>
+      <QueryClientProvider client={queryClient}>
+        <FilteredDataProvider>{children}</FilteredDataProvider>
+      </QueryClientProvider>
+    </MapProvider>
+  );
+};
+
+const FilteredDataProvider = ({ children }: { children: React.ReactNode }) => {
+  const { arboMap } = useMap();
+  const [state, dispatch] = useReducer(
+    (state: ArboStateType, action: ArboAction) =>
+      arboReducer(state, action, arboMap),
+    initialState
+  );
+  const dataQuery = useArboData();
+
+  useEffect(() => {
+    if (
+      state.filteredData.length === 0 &&
+      !state.dataFiltered &&
+      "data" in dataQuery &&
+      !!dataQuery.data &&
+      typeof dataQuery.data === "object" &&
+      "arbovirusEstimates" in dataQuery.data &&
+      Array.isArray(dataQuery.data.arbovirusEstimates) &&
+      dataQuery.data.arbovirusEstimates.length > 0
+    ) {
+      dispatch({
+        type: ArboActionType.INITIAL_DATA_FETCH,
+        payload: { data: dataQuery.data.arbovirusEstimates },
+      });
+    }
+  }, [dataQuery]);
+
+  return (
+    <ArboContext.Provider value={{ ...state, dispatch: dispatch }}>
+      {children}
+    </ArboContext.Provider>
+  );
+};
+
+export function Hydrate(props: HydrateProps) {
+  return <RQHydrate {...props} />;
+}
