+++ contexts/arbo-context/filter-update-steps/add-action-to-selected-filters-step.tsx
@@ -1,0 +1,18 @@
+import { HandleArboFilterUpdateInput, HandleArboFilterUpdateOutput } from "../arbo-filter-update-steps";
+
+export const addActionToSelectedFiltersStep = (
+  input: HandleArboFilterUpdateInput
+): HandleArboFilterUpdateOutput => {
+  const newSelectedFilters = {
+    ...input.state.selectedFilters,
+    [input.action.payload.filter]: input.action.payload.value,
+  }
+
+  return {
+    ...input,
+    state: {
+      ...input.state,
+      selectedFilters: newSelectedFilters
+    },
+  }
+};
