+++ package.json
@@ -36,6 +36,7 @@
     "eslint": "8.47.0",
     "eslint-config-next": "13.4.19",
     "export-to-csv": "^1.2.1",
+    "fp-ts": "^2.16.2",
     "graphql-request": "^6.1.0",
     "i18n-iso-countries": "^7.7.0",
     "lucide-react": "^0.268.0",
